refactor: Content retrieval
