+++ src/components/controllers/not-found.jsx
@@ -1,16 +1,10 @@
-import { useLanguage } from '../../lib/i18n';
-import { useContent } from '../../lib/use-resource';
-import { useTitle, useDescription } from './utils';
+import { useContent } from '../../lib/use-content';
 import Footer from '../footer';
 import { MarkdownRegion } from './markdown-region';
 import style from './style.module.css';
 
 export function NotFound() {
-	const [lang] = useLanguage();
-
-	const { html, meta } = useContent([lang, '404']);
-	useTitle(meta.title);
-	useDescription(meta.description);
+	const { html, meta } = useContent('404');
 
 	return (
 		<div class={style.page}>
