+++ src/components/controllers/repl-page.jsx
@@ -2,19 +2,15 @@ import { useLocation, useRoute } from 'preact-iso';
 import { Repl } from './repl';
 import { base64ToText } from './repl/query-encode.js';
 import { fetchExample } from './repl/examples';
-import { useContent, useResource } from '../../lib/use-resource';
-import { useTitle, useDescription } from './utils';
-import { useLanguage } from '../../lib/i18n';
+import { useResource } from '../../lib/use-resource';
+import { useContent } from '../../lib/use-content';
 
 import style from './repl/style.module.css';
 
 export default function ReplPage() {
 	const { query } = useRoute();
-	const [lang] = useLanguage();
 
-	const { meta } = useContent([lang, 'repl']);
-	useTitle(meta.title);
-	useDescription(meta.description);
+	useContent('repl');
 
 	const code = useResource(() => getInitialCode(query), [query]);
