+++ src/lib/use-content.js
@@ -1,6 +1,19 @@
 import { useEffect } from 'preact/hooks';
 
-import { createTitle } from '../../lib/page-title';
+import { createTitle } from './page-title';
+import { fetchContent } from './use-resource.js';
+
+/**
+ * @param {string} path
+ * @returns {{ html: string, meta: any }}
+ */
+export function useContent(path) {
+	const { html, meta } = fetchContent(path);
+	useTitle(meta.title);
+	useDescription(meta.description);
+
+	return { html, meta };
+}
 
 /**
  * Set `document.title`
